+++ plotly/src/contour.rs
@@ -445,7 +445,7 @@ where
     Y: Serialize + Default + Clone,
     Z: Serialize + Default + Clone,
 {
-    fn serialize(&self) -> String {
+    fn to_json(&self) -> String {
         serde_json::to_string(&self).unwrap()
     }
 }