+++ plotly/src/plot.rs
@@ -3,7 +3,9 @@ extern crate plotly_kaleido;
 
 use askama::Template;
 use dyn_clone::DynClone;
+use erased_serde::Serialize as ErasedSerialize;
 use rand::{thread_rng, Rng};
+use serde::Serialize;
 use std::env;
 use std::fs::File;
 use std::io::Write;
@@ -22,7 +24,7 @@ struct PlotlyJs;
 #[derive(Template)]
 #[template(path = "plot.html", escape = "none")]
 struct PlotTemplate<'a> {
-    plot_data: &'a str,
+    plot: &'a Plot,
     plotly_javascript: &'a str,
     remote_plotly_js: bool,
     export_image: bool,
@@ -34,14 +36,14 @@ struct PlotTemplate<'a> {
 #[derive(Template)]
 #[template(path = "inline_plot.html", escape = "none")]
 struct InlinePlotTemplate<'a> {
-    plot_data: &'a str,
+    plot: &'a Plot,
     plot_div_id: &'a str,
 }
 
 #[derive(Template)]
 #[template(path = "jupyter_notebook_plot.html", escape = "none")]
 struct JupyterNotebookPlotTemplate<'a> {
-    plot_data: &'a str,
+    plot: &'a Plot,
     plot_div_id: &'a str,
 }
 
@@ -56,11 +58,42 @@ pub enum ImageFormat {
 }
 
 /// A struct that implements `Trace` can be serialized to json format that is understood by Plotly.js.
-pub trait Trace: DynClone {
-    fn serialize(&self) -> String;
+pub trait Trace: DynClone + ErasedSerialize {
+    fn to_json(&self) -> String;
 }
 
 dyn_clone::clone_trait_object!(Trace);
+erased_serde::serialize_trait_object!(Trace);
+
+#[derive(Default, Serialize, Clone)]
+#[serde(transparent)]
+pub struct Traces {
+    traces: Vec<Box<dyn Trace>>,
+}
+
+impl Traces {
+    pub fn new() -> Self {
+        Self {
+            traces: Vec::with_capacity(1),
+        }
+    }
+
+    pub fn push(&mut self, trace: Box<dyn Trace>) {
+        self.traces.push(trace)
+    }
+
+    pub fn len(&self) -> usize {
+        self.traces.len()
+    }
+
+    pub fn iter(&self) -> std::slice::Iter<'_, Box<dyn Trace>> {
+        self.traces.iter()
+    }
+
+    pub fn to_json(&self) -> String {
+        serde_json::to_string(self).unwrap()
+    }
+}
 
 /// Plot is a container for structs that implement the `Trace` trait. Optionally a `Layout` can
 /// also be specified. Its function is to serialize `Trace`s and the `Layout` in html format and
@@ -71,7 +104,7 @@ dyn_clone::clone_trait_object!(Trace);
 /// ```
 /// extern crate plotly;
 /// use plotly::common::Mode;
-/// use plotly::{Plot, Scatter};
+/// use plotly::{Layout, Plot, Scatter};
 ///
 /// fn line_and_scatter_plot() {
 ///     let trace1 = Scatter::new(vec![1, 2, 3, 4], vec![10, 15, 13, 17])
@@ -86,6 +119,10 @@ dyn_clone::clone_trait_object!(Trace);
 ///     plot.add_trace(trace1);
 ///     plot.add_trace(trace2);
 ///     plot.add_trace(trace3);
+///
+///     let layout = Layout::new().title("<b>Line and Scatter Plot</b>".into());
+///     plot.set_layout(layout);
+///
 ///     plot.show();
 /// }
 ///
@@ -94,11 +131,14 @@ dyn_clone::clone_trait_object!(Trace);
 ///     Ok(())
 /// }
 /// ```
-#[derive(Default, Clone)]
+#[derive(Default, Serialize, Clone)]
 pub struct Plot {
-    traces: Vec<Box<dyn Trace>>,
-    layout: Option<Layout>,
-    configuration: Option<Configuration>,
+    #[serde(rename = "data")]
+    traces: Traces,
+    layout: Layout,
+    #[serde(rename = "config")]
+    configuration: Configuration,
+    #[serde(skip)]
     remote_plotly_js: bool,
 }
 
@@ -127,7 +167,7 @@ impl Plot {
     /// Create a new `Plot`.
     pub fn new() -> Plot {
         Plot {
-            traces: Vec::with_capacity(1),
+            traces: Traces::new(),
             remote_plotly_js: true,
             ..Default::default()
         }
@@ -153,12 +193,27 @@ impl Plot {
 
     /// Set the `Layout` to be used by `Plot`.
     pub fn set_layout(&mut self, layout: Layout) {
-        self.layout = Some(layout);
+        self.layout = layout;
     }
 
     /// Set the `Configuration` to be used by `Plot`.
     pub fn set_configuration(&mut self, configuration: Configuration) {
-        self.configuration = Some(configuration);
+        self.configuration = configuration;
+    }
+
+    /// Get the contained data elements.
+    pub fn data(&self) -> &Traces {
+        &self.traces
+    }
+
+    /// Get the layout specification of the plot.
+    pub fn layout(&self) -> &Layout {
+        &self.layout
+    }
+
+    /// Get the configuration specification of the plot.
+    pub fn configuration(&self) -> &Configuration {
+        &self.configuration
     }
 
     /// Renders the contents of the `Plot` and displays them in the system default browser.
@@ -310,10 +365,9 @@ impl Plot {
                 .collect::<Vec<u8>>(),
         )
         .unwrap();
-        let plot_data = self.render_plot_data();
 
         let tmpl = JupyterNotebookPlotTemplate {
-            plot_data: plot_data.as_str(),
+            plot: self,
             plot_div_id: plot_div_id.as_str(),
         };
         tmpl.render().unwrap()
@@ -381,34 +435,6 @@ impl Plot {
         templates.join(PLOTLY_JS)
     }
 
-    fn render_plot_data(&self) -> String {
-        let mut plot_data = String::new();
-        for (idx, trace) in self.traces.iter().enumerate() {
-            let s = trace.serialize();
-            plot_data.push_str(format!("var trace_{} = {};\n", idx, s).as_str());
-        }
-
-        plot_data.push_str("var data = [");
-        for idx in 0..self.traces.len() {
-            if idx != self.traces.len() - 1 {
-                plot_data.push_str(format!("trace_{},", idx).as_str());
-            } else {
-                plot_data.push_str(format!("trace_{}", idx).as_str());
-            }
-        }
-        plot_data.push_str("];\n");
-        let layout_data = match &self.layout {
-            Some(layout) => format!("var layout = {};", serde_json::to_string(layout).unwrap()),
-            None => {
-                let mut s = String::from("var layout = {");
-                s.push_str("};");
-                s
-            }
-        };
-        plot_data.push_str(layout_data.as_str());
-        plot_data
-    }
-
     fn render(
         &self,
         export_image: bool,
@@ -416,10 +442,9 @@ impl Plot {
         image_width: usize,
         image_height: usize,
     ) -> String {
-        let plot_data = self.render_plot_data();
         let plotly_js = PlotlyJs {}.render().unwrap();
         let tmpl = PlotTemplate {
-            plot_data: plot_data.as_str(),
+            plot: self,
             plotly_javascript: plotly_js.as_str(),
             remote_plotly_js: self.remote_plotly_js,
             export_image,
@@ -431,49 +456,15 @@ impl Plot {
     }
 
     fn render_inline(&self, plot_div_id: &str) -> String {
-        let plot_data = self.render_plot_data();
-
         let tmpl = InlinePlotTemplate {
-            plot_data: plot_data.as_str(),
+            plot: self,
             plot_div_id,
         };
         tmpl.render().unwrap()
     }
 
     pub fn to_json(&self) -> String {
-        let mut plot_data: Vec<String> = Vec::new();
-        for trace in self.traces.iter() {
-            let s = trace.serialize();
-            plot_data.push(s);
-        }
-
-        let mut json_data = String::new();
-        json_data.push_str(r#"{"data": ["#);
-
-        for (index, data) in plot_data.iter().enumerate() {
-            if index < plot_data.len() - 1 {
-                json_data.push_str(data);
-                json_data.push(',');
-            } else {
-                json_data.push_str(data);
-                json_data.push(']');
-            }
-        }
-
-        let layout_data = match &self.layout {
-            Some(layout) => serde_json::to_string(layout).unwrap(),
-            None => "{}".to_string(),
-        };
-        json_data.push_str(&format!(r#", "layout": {}"#, layout_data));
-
-        let configuration_data = match &self.configuration {
-            Some(configuration) => serde_json::to_string(configuration).unwrap(),
-            None => "{}".to_string(),
-        };
-        json_data.push_str(&format!(r#", "config": {}"#, configuration_data));
-
-        json_data.push('}');
-        json_data
+        serde_json::to_string(self).unwrap()
     }
 
     #[cfg(target_os = "linux")]
@@ -510,6 +501,8 @@ impl PartialEq for Plot {
 
 #[cfg(test)]
 mod tests {
+    use serde_json::{json, to_value};
+
     use super::*;
     use crate::Scatter;
 
@@ -518,13 +511,6 @@ mod tests {
         let mut plot = Plot::new();
         plot.add_trace(trace1);
         plot
-    }
-
-    #[test]
-    fn test_to_json() {
-        let plot = create_test_plot();
-        let plot_json = plot.to_json();
-        println!("{}", plot_json);
     }
 
     #[test]
@@ -554,6 +540,87 @@ mod tests {
     fn test_lab_display() {
         let plot = create_test_plot();
         plot.lab_display();
+    }
+
+    #[test]
+    fn test_plot_serialize_simple() {
+        let plot = create_test_plot();
+        let expected = json!({
+            "data": [
+                {
+                    "type": "scatter",
+                    "name": "trace1",
+                    "x": [0, 1, 2],
+                    "y": [6, 10, 2]
+                }
+            ],
+            "layout": {},
+            "config": {},
+        });
+
+        assert_eq!(to_value(plot).unwrap(), expected);
+    }
+
+    #[test]
+    fn test_plot_serialize_with_layout() {
+        let mut plot = create_test_plot();
+        let layout = Layout::new().title("Title".into());
+        plot.set_layout(layout);
+
+        let expected = json!({
+            "data": [
+                {
+                    "type": "scatter",
+                    "name": "trace1",
+                    "x": [0, 1, 2],
+                    "y": [6, 10, 2]
+                }
+            ],
+            "layout": {
+                "title": {
+                    "text": "Title"
+                }
+            },
+            "config": {},
+        });
+
+        assert_eq!(to_value(plot).unwrap(), expected);
+    }
+
+    #[test]
+    fn test_data_to_json() {
+        let plot = create_test_plot();
+        let expected = json!([
+            {
+                "type": "scatter",
+                "name": "trace1",
+                "x": [0, 1, 2],
+                "y": [6, 10, 2]
+            }
+        ]);
+
+        assert_eq!(to_value(plot.data()).unwrap(), expected);
+    }
+
+    #[test]
+    fn test_empty_layout_to_json() {
+        let plot = create_test_plot();
+        let expected = json!({});
+
+        assert_eq!(to_value(plot.layout()).unwrap(), expected);
+    }
+
+    #[test]
+    fn test_layout_to_json() {
+        let mut plot = create_test_plot();
+        let layout = Layout::new().title("TestTitle".into());
+        plot.set_layout(layout);
+
+        let expected = json!({
+            "title": {"text": "TestTitle"}
+        });
+
+        assert_eq!(to_value(plot.layout()).unwrap(), expected);
     }
 
     #[test]