+++ plotly/src/scatter_polar.rs
@@ -627,7 +627,7 @@ where
     Theta: Serialize + Clone + 'static,
     R: Serialize + Clone + 'static,
 {
-    fn serialize(&self) -> String {
+    fn to_json(&self) -> String {
         serde_json::to_string(&self).unwrap()
     }
 }