+++ plotly/src/surface.rs
@@ -477,7 +477,7 @@ where
     Y: Serialize + Clone,
     Z: Serialize + Clone,
 {
-    fn serialize(&self) -> String {
+    fn to_json(&self) -> String {
         serde_json::to_string(&self).unwrap()
     }
 }